refactor scattergl - separate calc and scene_update from index file
