+++ src/traces/scattergl/index.js
@@ -11,330 +11,21 @@
 var createScatter = require('regl-scatter2d');
 var createLine = require('regl-line2d');
 var createError = require('regl-error2d');
-var cluster = require('point-cluster');
 var Text = require('gl-text');
 
 var Registry = require('../../registry');
 var Lib = require('../../lib');
 var prepareRegl = require('../../lib/prepare_regl');
-var AxisIDs = require('../../plots/cartesian/axis_ids');
-var findExtremes = require('../../plots/cartesian/autorange').findExtremes;
 var Color = require('../../components/color');
 
 var subTypes = require('../scatter/subtypes');
-var scatterCalc = require('../scatter/calc');
-var calcMarkerSize = scatterCalc.calcMarkerSize;
-var calcAxisExpansion = scatterCalc.calcAxisExpansion;
-var setFirstScatter = scatterCalc.setFirstScatter;
-var calcColorscale = require('../scatter/colorscale_calc');
 var linkTraces = require('../scatter/link_traces');
 var getTraceColor = require('../scatter/get_trace_color');
 var fillText = Lib.fillText;
-var convert = require('./convert');
+var calc = require('./calc');
+var sceneUpdate = require('./scene_update');
 
-var BADNUM = require('../../constants/numerical').BADNUM;
-var TOO_MANY_POINTS = require('./constants').TOO_MANY_POINTS;
 var DESELECTDIM = require('../../constants/interactions').DESELECTDIM;
-
-function calc(gd, trace) {
-    var fullLayout = gd._fullLayout;
-    var xa = AxisIDs.getFromId(gd, trace.xaxis);
-    var ya = AxisIDs.getFromId(gd, trace.yaxis);
-    var subplot = fullLayout._plots[trace.xaxis + trace.yaxis];
-    var len = trace._length;
-    var hasTooManyPoints = len >= TOO_MANY_POINTS;
-    var len2 = len * 2;
-    var stash = {};
-    var i, xx, yy;
-
-    var x = xa.makeCalcdata(trace, 'x');
-    var y = ya.makeCalcdata(trace, 'y');
-
-    // we need hi-precision for scatter2d,
-    // regl-scatter2d uses NaNs for bad/missing values
-    var positions = new Array(len2);
-    for(i = 0; i < len; i++) {
-        xx = x[i];
-        yy = y[i];
-        positions[i * 2] = xx === BADNUM ? NaN : xx;
-        positions[i * 2 + 1] = yy === BADNUM ? NaN : yy;
-    }
-
-    if(xa.type === 'log') {
-        for(i = 0; i < len2; i += 2) {
-            positions[i] = xa.c2l(positions[i]);
-        }
-    }
-    if(ya.type === 'log') {
-        for(i = 1; i < len2; i += 2) {
-            positions[i] = ya.c2l(positions[i]);
-        }
-    }
-
-    // we don't build a tree for log axes since it takes long to convert log2px
-    // and it is also
-    if(hasTooManyPoints && (xa.type !== 'log' && ya.type !== 'log')) {
-        // FIXME: delegate this to webworker
-        stash.tree = cluster(positions);
-    } else {
-        var ids = stash.ids = new Array(len);
-        for(i = 0; i < len; i++) {
-            ids[i] = i;
-        }
-    }
-
-    // create scene options and scene
-    calcColorscale(gd, trace);
-    var opts = sceneOptions(gd, subplot, trace, positions, x, y);
-    var scene = sceneUpdate(gd, subplot);
-
-    // Reuse SVG scatter axis expansion routine.
-    // For graphs with very large number of points and array marker.size,
-    // use average marker size instead to speed things up.
-    setFirstScatter(fullLayout, trace);
-    var ppad;
-    if(!hasTooManyPoints) {
-        ppad = calcMarkerSize(trace, len);
-    } else if(opts.marker) {
-        ppad = 2 * (opts.marker.sizeAvg || Math.max(opts.marker.size, 3));
-    }
-    calcAxisExpansion(gd, trace, xa, ya, x, y, ppad);
-    if(opts.errorX) expandForErrorBars(trace, xa, opts.errorX);
-    if(opts.errorY) expandForErrorBars(trace, ya, opts.errorY);
-
-    // set flags to create scene renderers
-    if(opts.fill && !scene.fill2d) scene.fill2d = true;
-    if(opts.marker && !scene.scatter2d) scene.scatter2d = true;
-    if(opts.line && !scene.line2d) scene.line2d = true;
-    if((opts.errorX || opts.errorY) && !scene.error2d) scene.error2d = true;
-    if(opts.text && !scene.glText) scene.glText = true;
-
-    // FIXME: organize it in a more appropriate manner, probably in sceneOptions
-    // put point-cluster instance for optimized regl calc
-    if(opts.marker) {
-        opts.marker.snap = stash.tree || TOO_MANY_POINTS;
-    }
-
-    scene.lineOptions.push(opts.line);
-    scene.errorXOptions.push(opts.errorX);
-    scene.errorYOptions.push(opts.errorY);
-    scene.fillOptions.push(opts.fill);
-    scene.markerOptions.push(opts.marker);
-    scene.markerSelectedOptions.push(opts.markerSel);
-    scene.markerUnselectedOptions.push(opts.markerUnsel);
-    scene.textOptions.push(opts.text);
-    scene.textSelectedOptions.push(opts.textSel);
-    scene.textUnselectedOptions.push(opts.textUnsel);
-    scene.selectBatch.push([]);
-    scene.unselectBatch.push([]);
-
-    stash._scene = scene;
-    stash.index = scene.count;
-    stash.x = x;
-    stash.y = y;
-    stash.positions = positions;
-    scene.count++;
-
-    return [{x: false, y: false, t: stash, trace: trace}];
-}
-
-function expandForErrorBars(trace, ax, opts) {
-    var extremes = trace._extremes[ax._id];
-    var errExt = findExtremes(ax, opts._bnds, {padded: true});
-    extremes.min = extremes.min.concat(errExt.min);
-    extremes.max = extremes.max.concat(errExt.max);
-}
-
-function sceneOptions(gd, subplot, trace, positions, x, y) {
-    var opts = convert.style(gd, trace);
-
-    if(opts.marker) {
-        opts.marker.positions = positions;
-    }
-
-    if(opts.line && positions.length > 1) {
-        Lib.extendFlat(
-            opts.line,
-            convert.linePositions(gd, trace, positions)
-        );
-    }
-
-    if(opts.errorX || opts.errorY) {
-        var errors = convert.errorBarPositions(gd, trace, positions, x, y);
-
-        if(opts.errorX) {
-            Lib.extendFlat(opts.errorX, errors.x);
-        }
-        if(opts.errorY) {
-            Lib.extendFlat(opts.errorY, errors.y);
-        }
-    }
-
-    if(opts.text) {
-        Lib.extendFlat(
-            opts.text,
-            {positions: positions},
-            convert.textPosition(gd, trace, opts.text, opts.marker)
-        );
-        Lib.extendFlat(
-            opts.textSel,
-            {positions: positions},
-            convert.textPosition(gd, trace, opts.text, opts.markerSel)
-        );
-        Lib.extendFlat(
-            opts.textUnsel,
-            {positions: positions},
-            convert.textPosition(gd, trace, opts.text, opts.markerUnsel)
-        );
-    }
-
-    return opts;
-}
-
-// make sure scene exists on subplot, return it
-function sceneUpdate(gd, subplot) {
-    var scene = subplot._scene;
-
-    var resetOpts = {
-        // number of traces in subplot, since scene:subplot -> 1:1
-        count: 0,
-        // whether scene requires init hook in plot call (dirty plot call)
-        dirty: true,
-        // last used options
-        lineOptions: [],
-        fillOptions: [],
-        markerOptions: [],
-        markerSelectedOptions: [],
-        markerUnselectedOptions: [],
-        errorXOptions: [],
-        errorYOptions: [],
-        textOptions: [],
-        textSelectedOptions: [],
-        textUnselectedOptions: [],
-        // selection batches
-        selectBatch: [],
-        unselectBatch: []
-    };
-
-    // regl- component stubs, initialized in dirty plot call
-    var initOpts = {
-        fill2d: false,
-        scatter2d: false,
-        error2d: false,
-        line2d: false,
-        glText: false,
-        select2d: false
-    };
-
-    if(!subplot._scene) {
-        scene = subplot._scene = {};
-
-        scene.init = function init() {
-            Lib.extendFlat(scene, initOpts, resetOpts);
-        };
-
-        scene.init();
-
-        // apply new option to all regl components (used on drag)
-        scene.update = function update(opt) {
-            var opts = Lib.repeat(opt, scene.count);
-
-            if(scene.fill2d) scene.fill2d.update(opts);
-            if(scene.scatter2d) scene.scatter2d.update(opts);
-            if(scene.line2d) scene.line2d.update(opts);
-            if(scene.error2d) scene.error2d.update(opts.concat(opts));
-            if(scene.select2d) scene.select2d.update(opts);
-            if(scene.glText) {
-                for(var i = 0; i < scene.count; i++) {
-                    scene.glText[i].update(opt);
-                }
-            }
-        };
-
-        // draw traces in proper order
-        scene.draw = function draw() {
-            var count = scene.count;
-            var fill2d = scene.fill2d;
-            var error2d = scene.error2d;
-            var line2d = scene.line2d;
-            var scatter2d = scene.scatter2d;
-            var glText = scene.glText;
-            var select2d = scene.select2d;
-            var selectBatch = scene.selectBatch;
-            var unselectBatch = scene.unselectBatch;
-
-            for(var i = 0; i < count; i++) {
-                if(fill2d && scene.fillOrder[i]) {
-                    fill2d.draw(scene.fillOrder[i]);
-                }
-                if(line2d && scene.lineOptions[i]) {
-                    line2d.draw(i);
-                }
-                if(error2d) {
-                    if(scene.errorXOptions[i]) error2d.draw(i);
-                    if(scene.errorYOptions[i]) error2d.draw(i + count);
-                }
-                if(scatter2d && scene.markerOptions[i]) {
-                    if(unselectBatch[i].length) {
-                        var arg = Lib.repeat([], scene.count);
-                        arg[i] = unselectBatch[i];
-                        scatter2d.draw(arg);
-                    } else if(!selectBatch[i].length) {
-                        scatter2d.draw(i);
-                    }
-                }
-                if(glText[i] && scene.textOptions[i]) {
-                    glText[i].render();
-                }
-            }
-
-            if(select2d) {
-                select2d.draw(selectBatch);
-            }
-
-            scene.dirty = false;
-        };
-
-        // remove scene resources
-        scene.destroy = function destroy() {
-            if(scene.fill2d && scene.fill2d.destroy) scene.fill2d.destroy();
-            if(scene.scatter2d && scene.scatter2d.destroy) scene.scatter2d.destroy();
-            if(scene.error2d && scene.error2d.destroy) scene.error2d.destroy();
-            if(scene.line2d && scene.line2d.destroy) scene.line2d.destroy();
-            if(scene.select2d && scene.select2d.destroy) scene.select2d.destroy();
-            if(scene.glText) {
-                scene.glText.forEach(function(text) {
-                    if(text.destroy) text.destroy();
-                });
-            }
-
-            scene.lineOptions = null;
-            scene.fillOptions = null;
-            scene.markerOptions = null;
-            scene.markerSelectedOptions = null;
-            scene.markerUnselectedOptions = null;
-            scene.errorXOptions = null;
-            scene.errorYOptions = null;
-            scene.textOptions = null;
-            scene.textSelectedOptions = null;
-            scene.textUnselectedOptions = null;
-
-            scene.selectBatch = null;
-            scene.unselectBatch = null;
-
-            // we can't just delete _scene, because `destroy` is called in the
-            // middle of supplyDefaults, before relinkPrivateKeys which will put it back.
-            subplot._scene = null;
-        };
-    }
-
-    // in case if we have scene from the last calc - reset data
-    if(!scene.dirty) {
-        Lib.extendFlat(scene, resetOpts);
-    }
-
-    return scene;
-}
 
 function getViewport(fullLayout, xaxis, yaxis) {
     var gs = fullLayout._size;
